Render PluginLoadState errors in the UI and allow parent components to override error UI (#4192)
Render more helpful error messages for cases where the TensorBoard server is not responsive or has an error. Allow other flavors of TensorBoard to customize these errors.

In https://github.com/tensorflow/tensorboard/pull/4186 we improved error handling for `data/runs`, `data/plugins_listing` and `data/environment` HTTP requests. Errors from those requests are translated into a `failureCode` field in the `pluginsListLoad` state.

In this PR we render those errors in the `plugins` component, given that we want to render the errors in a similar way to the errors the component already renders.

We also provide a mechanism for parent component to specify custom error templates. This will allow other flavors of TensorBoard to customize the error messages.